# evmrs: use little endian for u256 (and stack)