--- a/rust/fuzz/fuzz_targets/evmc_execute.rs
+++ b/rust/fuzz/fuzz_targets/evmc_execute.rs
@@ -24,7 +24,7 @@ struct InterpreterArgs<'a> {
     code: &'a [u8],
 }
 
-impl<'a> Debug for InterpreterArgs<'a> {
+impl Debug for InterpreterArgs<'_> {
     fn fmt(&self, f: &mut std::fmt::Formatter<'_>) -> std::fmt::Result {
         f.debug_struct("InterpreterArgs")
             .field("host", &self.host)