This PR changes the u256 type to hold a U256 instead of [u8; 32]. Before it was expected that the bytes in [u8; 32] are in big endian, while U256 is little endian. With this change, it is no longer necessary to convert to little endian and back when doing arithmetic operations. However, call data load, mstore, mload and push have to do the conversion now. However the benchmarks show that this is faster. Further more the conversion for push can be done in the code analysis (for feature fn-ptr-conversion-expanded-dispatch), which gets cached.


```
                   │ 2024-11-27T18:12#1207189#20-main/evmrs#performance │ 2024-11-29T08:27#5f23203#10/evmrs#performance │
                   │                       sec/op                       │       sec/op         vs base                  │
StaticOverhead/1/                                           2.664µ ± 0%           2.603µ ± 1%  -2.31% (p=0.000 n=20+10)
Inc/1/                                                      5.157µ ± 0%           5.247µ ± 1%  +1.76% (p=0.000 n=20+10)
Inc/10/                                                     5.170µ ± 0%           5.284µ ± 0%  +2.21% (p=0.000 n=20+10)
Fib/1/                                                      4.585µ ± 0%           4.611µ ± 0%  +0.58% (p=0.013 n=20+10)
Fib/5/                                                      14.18µ ± 0%           14.28µ ± 1%  +0.66% (p=0.001 n=20+10)
Fib/10/                                                     129.5µ ± 0%           129.7µ ± 0%       ~ (p=0.109 n=20+10)
Fib/15/                                                     1.231m ± 0%           1.229m ± 0%  -0.14% (p=0.000 n=20+10)
Fib/20/                                                     13.26m ± 0%           13.18m ± 0%  -0.60% (p=0.000 n=20+10)
Sha3/1/                                                     2.874µ ± 0%           2.867µ ± 1%       ~ (p=0.454 n=20+10)
Sha3/10/                                                    4.319µ ± 0%           4.252µ ± 1%  -1.56% (p=0.000 n=20+10)
Sha3/100/                                                   18.28µ ± 0%           17.28µ ± 0%  -5.48% (p=0.000 n=20+10)
Sha3/1000/                                                  188.8µ ± 0%           183.8µ ± 0%  -2.64% (p=0.000 n=20+10)
Arith/1/                                                    5.837µ ± 0%           5.926µ ± 1%  +1.52% (p=0.000 n=20+10)
Arith/10/                                                   12.09µ ± 0%           11.38µ ± 0%  -5.89% (p=0.000 n=20+10)
Arith/100/                                                  72.74µ ± 0%           68.72µ ± 0%  -5.52% (p=0.000 n=20+10)
Arith/280/                                                  212.4µ ± 0%           200.5µ ± 0%  -5.61% (p=0.000 n=20+10)
Memory/1/                                                   8.161µ ± 1%           8.165µ ± 1%       ~ (p=0.820 n=20+10)
Memory/10/                                                  13.71µ ± 0%           13.37µ ± 3%       ~ (p=0.262 n=20+10)
Memory/100/                                                 72.81µ ± 0%           66.81µ ± 0%  -8.24% (p=0.000 n=20+10)
Memory/1000/                                                616.7µ ± 0%           571.5µ ± 0%  -7.33% (p=0.000 n=20+10)
Memory/10000/                                               5.888m ± 0%           5.377m ± 0%  -8.69% (p=0.000 n=20+10)
Analysis/jumpdest/                                          2.578µ ± 1%           2.550µ ± 1%  -1.11% (p=0.003 n=20+10)
Analysis/stop/                                              2.572µ ± 0%           2.554µ ± 1%  -0.70% (p=0.001 n=20+10)
Analysis/push1/                                             2.558µ ± 0%           2.548µ ± 0%  -0.39% (p=0.019 n=20+10)
Analysis/push32/                                            2.559µ ± 0%           2.541µ ± 1%  -0.72% (p=0.008 n=20+10)
geomean                                                     26.96µ                26.38µ       -2.16%
```